requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-extensions==2.1.6
django-taggit==1.1.0
djangorestframework==3.9.4
drf-yasg==1.15.0
entrypoints==0.3
flake8==3.7.7
idna==2.8
inflection==0.3.1
isort==4.3.19
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.24.3
wrapt==1.11.1
